+++ src/infrastructure/prun-ui/buffers.ts
@@ -6,6 +6,7 @@ import { getPrunId } from '@src/infrastructure/prun-ui/attributes';
 import { watchUntil } from '@src/utils/watch';
 import { isEmpty } from 'ts-extras';
 import { dispatchClientPrunMessage } from '@src/infrastructure/prun-api/prun-api-listener';
+import { clamp } from '@src/utils/clamp';
 
 let isBusy = false;
 const pendingResolvers: (() => void)[] = [];
@@ -61,7 +62,7 @@ function releaseSlot() {
   }
 }
 
-async function captureLastWindow(fullCommand: string, options?: ShowBufferOptions) {
+async function captureLastWindow(command: string, options?: ShowBufferOptions) {
   const windows = _$$(document, C.Window.window);
   if (isEmpty(windows)) {
     return;
@@ -73,7 +74,7 @@ async function captureLastWindow(fullCommand: string, options?: ShowBufferOption
     return;
   }
   if (!(options?.autoSubmit ?? true)) {
-    changeInputValue(input, fullCommand);
+    changeInputValue(input, command);
     return;
   }
   window.classList.add(css.hidden);
@@ -92,11 +93,11 @@ async function captureLastWindow(fullCommand: string, options?: ShowBufferOption
     messageType: 'UI_TILES_CHANGE_COMMAND',
     payload: {
       id: id,
-      newCommand: fullCommand,
+      newCommand: command,
     },
   };
   if (!dispatchClientPrunMessage(message)) {
-    changeInputValue(input, fullCommand);
+    changeInputValue(input, command);
     await sleep(0);
     form.requestSubmit();
   }
@@ -107,21 +108,6 @@ async function captureLastWindow(fullCommand: string, options?: ShowBufferOption
   ]);
   if (!options?.autoClose) {
     window.classList.remove(css.hidden);
-    const command = fullCommand.split(' ')[0];
-    const defaultSize = defaultBufferSizes[command];
-    if (defaultSize) {
-      const message = {
-        messageType: 'UI_WINDOWS_UPDATE_SIZE',
-        payload: {
-          id: id,
-          size: {
-            width: defaultSize[0],
-            height: defaultSize[1],
-          },
-        },
-      };
-      dispatchClientPrunMessage(message);
-    }
     return;
   }
   void closeWhenDone(window, options);
@@ -141,51 +127,15 @@ async function closeWhenDone(window: HTMLDivElement, options?: ShowBufferOptions
   await new Promise<void>(resolve => onNodeDisconnected(window, resolve));
 }
 
-const defaultBufferSizes = {
-  ADM: [380, 550],
-  BBC: [500, 450],
-  BLU: [550, 600],
-  BS: [610, 300],
-  BSC: [550, 620],
-  BTF: [570, 700],
-  BUI: [500, 400],
-  COGC: [500, 580],
-  CONT: [600, 400],
-  CONTD: [450, 550],
-  CONTS: [550, 300],
-  CORPARC: [350, 550],
-  CORPNP: [450, 430],
-  CORPP: [460, 640],
-  CX: [550, 600],
-  CXL: [600, 180],
-  CXM: [625, 300],
-  CXOS: [750, 300],
-  CXPO: [450, 310],
-  FLT: [650, 180],
-  GOV: [470, 550],
-  HQ: [450, 600],
-  INV: [530, 250],
-  LEAD: [700, 400],
-  LM: [500, 580],
-  LMA: [425, 370],
-  LMOS: [700, 420],
-  LMP: [450, 500],
-  MAT: [500, 400],
-  MOTS: [600, 450],
-  MU: [512, 512],
-  NOTS: [425, 625],
-  PLI: [450, 600],
-  POPI: [550, 300],
-  POPID: [460, 500],
-  POPR: [515, 400],
-  PROD: [400, 500],
-  PRODCO: [415, 600],
-  PRODQ: [650, 300],
-  SHP: [450, 450],
-  SHY: [450, 450],
-  STEAM: [300, 450],
-  STNS: [400, 280],
-  SYSI: [600, 600],
-  WAR: [400, 580],
-  WF: [710, 300],
-};
+export function setBufferSize(id: string, width: number, height: number) {
+  dispatchClientPrunMessage({
+    messageType: 'UI_WINDOWS_UPDATE_SIZE',
+    payload: {
+      id: id,
+      size: {
+        width: clamp(width, 100, document.body.clientWidth - 50),
+        height: clamp(height, 50, document.body.clientHeight - 50),
+      },
+    },
+  });
+}