+++ src/types/globals.d.ts
@@ -1,5 +1,6 @@
 interface PrunTile {
   id: string;
+  container: HTMLElement;
   frame: HTMLDivElement;
   anchor: HTMLDivElement;
   docked: boolean;